Skeletal animation is now fully implemented

+++ FbxConverter/MeshExporter.h
@@ -52,13 +52,13 @@ struct SkinnedVertex {
 	glm::vec2 uv;
 
 	int joint_index_0 = 0;
-	int joint_weight_0 = 0.0f;
+	float joint_weight_0 = 0.0f;
 	int joint_index_1 = 0;
-	int joint_weight_1 = 0.0f;
+	float joint_weight_1 = 0.0f;
 	int joint_index_2 = 0;
-	int joint_weight_2 = 0.0f;
+	float joint_weight_2 = 0.0f;
 	int joint_index_3 = 0;
-	int joint_weight_3 = 0.0f;
+	float joint_weight_3 = 0.0f;
 
 
 	// bool operator<(const Vertex& v2) const{
